update interface imports examples and docs
