--- examples/tutorial/src/interfaces/ICategory.d.ts
@@ -1,4 +1,0 @@
-export interface ICategory {
-    id: string;
-    title: string;
-}
