--- examples/tutorial/src/interfaces/IPost.d.ts
@@ -1,8 +1,0 @@
-import { ICategory } from "./ICategory";
-
-export interface IPost {
-    title: string;
-    status: "published" | "draft";
-    category: ICategory;
-    createdAt: string;
-}
