+++ examples/tutorial/src/pages/posts/list.tsx
@@ -15,7 +15,7 @@ import {
     useSelect,
 } from "@pankod/refine";
 
-import { IPost, ICategory } from "../../interfaces";
+import { IPost, ICategory } from "interfaces";
 
 export const PostList = (props: IResourceComponentsProps) => {
     const { tableProps } = useTable<IPost>();
