+++ examples/tutorial/src/pages/posts/show.tsx
@@ -1,5 +1,5 @@
 import { Show, useShow, Typography, Tag, useOne } from "@pankod/refine";
-import { IPost, ICategory } from "../../interfaces";
+import { IPost, ICategory } from "interfaces";
 
 const { Title, Text } = Typography;
