--- examples/useModalForm/src/interfaces/IPost.d.ts
@@ -1,8 +1,0 @@
-import { ICategory } from "./";
-export interface IPost {
-    id: string;
-    title: string;
-    content: string;
-    status: "published" | "draft";
-    category: ICategory;
-}
