+++ examples/useModalForm/src/interfaces/index.d.ts
@@ -1,2 +1,12 @@
-export { ICategory } from "./ICategory";
-export { IPost } from "./IPost";
+export interface ICategory {
+    id: string;
+    title: string;
+}
+
+export interface IPost {
+    id: string;
+    title: string;
+    content: string;
+    status: "published" | "draft";
+    category: ICategory;
+}
